+++ flutter_google_street_view/lib/flutter_google_street_view_web.dart
@@ -2,11 +2,13 @@ import 'dart:async';
 import 'dart:html';
 import 'dart:math';
 
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_google_street_view/src/web/convert.dart';
-import 'package:flutter_google_street_view/src/web/street_view_state.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 import 'package:google_maps/google_maps.dart' as gmaps;
+import 'package:google_maps/google_maps.dart';
 import 'package:kotlin_scope_function/kotlin_scope_function.dart';
+import 'package:flutter_google_street_view/src/web/shims/dart_ui.dart' as ui;
 
 part 'package:flutter_google_street_view/src/web/plugin.dart';