+++ street_view_platform_interface/lib/src/type/latLng.dart
@@ -1,0 +1,49 @@
+import 'dart:ui' show hashValues;
+
+class LatLng {
+  /// Creates a geographical location specified in degrees [latitude] and
+  /// [longitude].
+  ///
+  /// The latitude is clamped to the inclusive interval from -90.0 to +90.0.
+  ///
+  /// The longitude is normalized to the half-open interval from -180.0
+  /// (inclusive) to +180.0 (exclusive)
+  const LatLng(double latitude, double longitude)
+      : assert(latitude != null),
+        assert(longitude != null),
+        latitude =
+            (latitude < -90.0 ? -90.0 : (90.0 < latitude ? 90.0 : latitude)),
+        longitude = (longitude + 180.0) % 360.0 - 180.0;
+
+  /// The latitude in degrees between -90.0 and 90.0, both inclusive.
+  final double latitude;
+
+  /// The longitude in degrees between -180.0 (inclusive) and 180.0 (exclusive).
+  final double longitude;
+
+  /// Converts this object to something serializable in JSON.
+  Object toJson() {
+    return <double>[latitude, longitude];
+  }
+
+  /// Initialize a LatLng from an \[lat, lng\] array.
+  static LatLng? fromJson(Object? json) {
+    if (json == null) {
+      return null;
+    }
+    assert(json is List && json.length == 2);
+    final list = json as List;
+    return LatLng(list[0], list[1]);
+  }
+
+  @override
+  String toString() => '$runtimeType($latitude, $longitude)';
+
+  @override
+  bool operator ==(Object o) {
+    return o is LatLng && o.latitude == latitude && o.longitude == longitude;
+  }
+
+  @override
+  int get hashCode => hashValues(latitude, longitude);
+}