+++ street_view_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+* `Web`, fine tune for [billing](https://developers.google.com/maps/documentation/javascript/usage-and-billing#dynamic-street-view) of Street View panorama. 
+
 ## 1.0.1
 
 * Remove dependency `google_maps_flutter_platform_interface`